% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_debiased_scores_dense}
\alias{cpp_debiased_scores_dense}
\title{Debiased target scores for general precision directions}
\usage{
cpp_debiased_scores_dense(X, y, beta_full, target_point, directions)
}
\description{
Debiased target scores for general precision directions
}
\keyword{internal}
