% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_debiased_scores_diagonal}
\alias{cpp_debiased_scores_diagonal}
\title{Debiased coordinate scores for a diagonal precision model}
\usage{
cpp_debiased_scores_diagonal(
  X,
  y,
  beta_full,
  targets,
  direction_scale,
  target_point
)
}
\description{
Debiased coordinate scores for a diagonal precision model
}
\keyword{internal}
