% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_multiplier_max}
\alias{cpp_multiplier_max}
\title{Multiplier-bootstrap maxima from centered influence scores}
\usage{
cpp_multiplier_max(
  psi_centered,
  standard_error,
  multipliers,
  block_size = 256L
)
}
\description{
Multiplier-bootstrap maxima from centered influence scores
}
\keyword{internal}
