% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_operator_norm}
\alias{cpp_operator_norm}
\title{RcppArmadillo operator norm for the stacked data/restriction operator}
\usage{
cpp_operator_norm(X, A, max_iter = 200L, tol = 1e-09)
}
\description{
RcppArmadillo operator norm for the stacked data/restriction operator
}
\keyword{internal}
