% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cpp_sqrt_lasso_pd}
\alias{cpp_sqrt_lasso_pd}
\title{RcppArmadillo primal-dual square-root lasso solver}
\usage{
cpp_sqrt_lasso_pd(
  X,
  y,
  lambda,
  A,
  target,
  penalty_factor,
  beta_init,
  max_iter = 30000L,
  tol = 5e-06,
  check_every = 50L,
  verbose = FALSE,
  operator_norm_value = NA_real_
)
}
\description{
RcppArmadillo primal-dual square-root lasso solver
}
\keyword{internal}
